Combine integration test struct into themes
