+++ test/integration/proxy-test.js
@@ -19,8 +19,12 @@
 // THE SOFTWARE.
 'use strict';
 
+var _ = require('underscore');
 var after = require('after');
+var jsonBody = require('body/json');
 var test = require('tape');
+var TimeMock = require('time-mock');
+var tryIt = require('tryit');
 
 var allocCluster = require('../lib/alloc-cluster.js');
 var strHead = require('../../lib/request-proxy/util.js').strHead;
@@ -33,6 +37,106 @@ function scheduleTicker(ringpop, onScheduled) {
         onScheduled();
     };
 }
+
+test('handleOrProxy() returns true for me', function t(assert) {
+    var cluster = allocCluster(function onReady() {
+        var keyOne = cluster.keys.one;
+
+        assert.equal(cluster.one.handleOrProxy(keyOne), true);
+
+        cluster.destroy();
+        assert.end();
+    });
+});
+
+test('handleOrProxy() proxies for not me', function t(assert) {
+    var cluster = allocCluster(function onReady() {
+        var keyTwo = cluster.keys.two;
+
+        cluster.request({
+            key: keyTwo,
+            host: 'one',
+            json: { hello: true }
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.statusCode, 200);
+            assert.equal(resp.body.host, 'two');
+            assert.equal(resp.body.payload.hello, true);
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+});
+
+test('handleOrProxyAll() proxies and handles locally', function t(assert) {
+    var handlerCallCounts = {};
+    var cluster = allocCluster({
+        createHandler: createServerHandler
+    }, function onReady() {
+        var k = cluster.keys;
+        var keys = [k.one, k.two, k.two, k.three];
+
+        cluster.requestAll({
+            keys: keys,
+            host: 'one',
+            json: { hello: true }
+        }, onResponses);
+    });
+
+    function onResponses(err, responses) {
+        assert.ifError(err);
+
+        assert.equal(responses.length, 3);
+        responses.forEach(function(data) {
+            assert.equal(data.res.statusCode, 200);
+            tryIt(function parse() {
+                var body = JSON.parse(data.res.body);
+                assert.equal(body.payload.hello, true);
+            }, assert.ifError);
+        });
+        assert.equal(handlerCallCounts.one, 1);
+        assert.equal(handlerCallCounts.two, 1);
+        assert.equal(handlerCallCounts.three, 1);
+
+        cluster.destroy();
+        assert.end();
+    }
+
+    function createServerHandler(name) {
+        return function serverHandle(req, res) {
+            if (handlerCallCounts[name]) {
+                handlerCallCounts[name]++;
+            } else {
+                handlerCallCounts[name] = 1;
+            }
+
+            if (req.headers['content-type'] === 'application/json') {
+                jsonBody(req, {cache: true}, onBody);
+            } else {
+                onBody(null, undefined);
+            }
+
+            function onBody(err, result) {
+                if (err) {
+                    res.statusCode = 500;
+                    return res.end(err.message);
+                }
+
+                res.statusCode = 200;
+                res.end(JSON.stringify({
+                    host: name,
+                    url: req.url,
+                    headers: req.headers,
+                    method: req.method,
+                    httpVersion: req.httpVersion,
+                    payload: result
+                }));
+            }
+        };
+    }
+});
 
 test('can proxyReq() to someone', function t(assert) {
     var cluster = allocCluster(function onReady() {
@@ -495,6 +599,324 @@ test('reroutes retry to remote', function t(assert) {
         }, function onRequest(err, res) {
             assert.ifError(err, 'no error occurs');
             assert.equal(res.body, 'rerouted', 'response from rerouted request is correct');
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+});
+
+// basic tests
+test('can serialize url', function t(assert) {
+    var cluster = allocCluster(function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two,
+            url: '/foo', json: true
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.body.url, '/foo');
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+});
+
+test('can serialize headers', function t(assert) {
+    var cluster = allocCluster(function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two,
+            headers: { 'foo': 'bar' }, json: true
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.body.headers.foo, 'bar');
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+});
+
+test('can serialize method', function t(assert) {
+    var cluster = allocCluster(function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two,
+            method: 'POST', json: true
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.body.method, 'POST');
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+});
+
+test('can serialize httpVersion', function t(assert) {
+    var cluster = allocCluster(function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two,
+            httpVersion: '1.1', json: true
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.body.httpVersion, '1.1');
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+});
+
+test('will timeout after default timeout', function t(assert) {
+    var timers = TimeMock(Date.now());
+    var cluster = allocCluster({
+        timers: timers,
+        createHandler: function createHandler() {
+            return function handle() {
+                // do nothing to timeout
+            };
+        },
+        requestProxyMaxRetries: 0
+    }, function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two,
+            json: { hello: true }
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.statusCode, 500);
+            // timeout non deterministically times the cb out
+            // or closes the TCP socket.
+            assert.ok(resp.body === 'timed out' ||
+                resp.body === 'socket closed');
+
+            cluster.destroy();
+            assert.end();
+        });
+
+        setTimeout(function onTick() {
+            timers.advance(32000);
+        }, 50);
+    });
+});
+
+test('can serialize body', function t(assert) {
+    var cluster = allocCluster(function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two,
+            json: { hello: true }
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.statusCode, 200);
+            assert.equal(resp.body.payload.hello, true);
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+});
+
+test('can serialize response statusCode', function t(assert) {
+    var cluster = allocCluster({
+         createHandler: function createHandler() {
+            return function handle(req, res) {
+                res.statusCode = 404;
+                res.end();
+            };
+        }
+    }, function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.statusCode, 404);
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+});
+
+test('can serialize response headers', function t(assert) {
+    var cluster = allocCluster({
+         createHandler: function createHandler() {
+            return function handle(req, res) {
+                res.setHeader('X-Foo', 'bar');
+                res.end();
+            };
+        }
+    }, function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.statusCode, 200);
+            assert.equal(resp.headers['x-foo'], 'bar');
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+});
+
+test('can serialize response body', function t(assert) {
+    var cluster = allocCluster({
+         createHandler: function createHandler() {
+            return function handle(req, res) {
+                res.end('hello');
+            };
+        }
+    }, function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.statusCode, 200);
+            assert.equal(resp.body, 'hello');
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+});
+
+// new features
+test('can handle errors differently');
+test('adds forwarding header');
+test('does not handle MockResponse errors');
+test('checks the checksum for response');
+test('can send back a close event');
+
+// lack of coverage.
+test('custom timeouts', function t(assert) {
+    var timers = TimeMock(Date.now());
+    var cluster = allocCluster({
+        timers: timers,
+        createHandler: function createHandler() {
+            return function handle() {
+                // do nothing to timeout
+            };
+        },
+        requestProxyMaxRetries: 0
+    }, function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two,
+            json: { hello: true },
+            timeout: 1500
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.statusCode, 500);
+            assert.equal(resp.body, 'timed out');
+
+            cluster.destroy();
+            assert.end();
+        });
+
+        setTimeout(function onTick() {
+            timers.advance(3200);
+        }, 50);
+    });
+});
+
+test('handle body failures', function t(assert) {
+    var cluster = allocCluster(function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two,
+            body: '1234567890',
+            headers: {
+                'content-length': '5'
+            }
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.statusCode, 400);
+            assert.equal(resp.body,
+                'request size did not match content length');
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+});
+
+test('non json head is ok', function t(assert) {
+    var cluster = allocCluster(function onReady() {
+        var two = cluster.two;
+
+        two.channel.register('/proxy/req', fakeProxyReq);
+
+        cluster.request({
+            host: 'one', key: cluster.keys.two
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.statusCode, 200);
+
+            cluster.destroy();
+            assert.end();
+        });
+    });
+
+    function fakeProxyReq(arg1, arg2, hostInfo, cb) {
+        cb(null, 'fake-text', '');
+    }
+});
+
+test('handle tchannel failures', function t(assert) {
+    var cluster = allocCluster({
+        createHandler: function createHandler() {
+            return function handle() {
+                // do nothing to timeout
+            };
+        },
+        requestProxyMaxRetries: 0
+    }, function onReady() {
+        cluster.request({
+            host: 'one', key: cluster.keys.two
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.statusCode, 500);
+            assert.equal(resp.body, 'socket closed');
+
+            cluster.destroy();
+            assert.end();
+        });
+
+        // Reach into the the first TChannel and forcibly
+        // destroy its open TCP connection
+        setTimeout(function onTimer() {
+            var name = cluster.two.channel.name;
+            cluster.one.channel.getPeer(name).socket.destroy();
+        }, 100);
+    });
+});
+
+test('handles checksum failures', function t(assert) {
+    var ringpopOpts = {
+        requestProxyMaxRetries: 0
+    };
+
+    var cluster = allocCluster(ringpopOpts, function onReady() {
+        cluster.two.ring.addServer('localhost:9999');
+
+        cluster.request({
+            host: 'one', key: cluster.keys.two
+        }, function onResponse(err, resp) {
+            assert.ifError(err);
+
+            assert.equal(resp.statusCode, 500);
+            assert.ok(resp.body.indexOf(
+                'Expected the remote checksum to match local checksum') >= 0);
 
             cluster.destroy();
             assert.end();
