+++ test/integration/ring-test.js
